backend/requirements.txt
django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==11.1.0 